Announcement web

+++ server/DBorm/DBorm.js
@@ -139,36 +139,38 @@ UsersChoresTypes.belongsTo(Users, {
 });
 
 
-
-sequelize.sync({force: true})
-    .then(() => {
-        Users.create({
-            userId: '1',
-            fullname: 'מנהל מערכת',
-            password: 'Admin123',
-            email: 'admin@gamil.com',
-            mailbox: 1,
-            cellphone: '0123456789',
-            phone: '0123456789',
-            bornDate: new Date('1992-11-25'),
-            active: true,
-        })
-            .then(user => {
-                ServiceProviders.create({
-                    serviceProviderId: 1,
-                    userId: user.userId,
-                    role: 'Admin',
-                    operationTime: 'all time',
-                    phoneNumber: '0123456789',
-                    appointmentWayType: 'Admin',
-                    subjects:"[\"הכל\"]",
-                    active: true,
-                })
+if(process.dbMode === "dev"){
+    sequelize.sync({force: true})
+        .then(() => {
+            Users.create({
+                userId: '1',
+                fullname: 'מנהל מערכת',
+                password: 'Admin123',
+                email: 'admin@gamil.com',
+                mailbox: 1,
+                cellphone: '0123456789',
+                phone: '0123456789',
+                bornDate: new Date('1992-11-25'),
+                active: true,
             })
-            .then(
-                console.log(`Database & tables created!`)
-            )
-    });
+                .then(user => {
+                    ServiceProviders.create({
+                        serviceProviderId: 1,
+                        userId: user.userId,
+                        role: 'Admin',
+                        operationTime: 'all time',
+                        phoneNumber: '0123456789',
+                        appointmentWayType: 'all',
+                        subjects:"[\"הכל\"]",
+                        active: true,
+                    })
+                })
+                .then(
+                    console.log(`Database & tables created!`)
+                )
+        });
+}
+
 
     
 module.exports = {
